Working unit tests

+++ src/modelconfig.cpp
@@ -480,7 +480,7 @@ Status ModelConfig::parseNode(const rapidjson::Value& v) {
                 } else {
                     for (auto& sh : s.value.GetArray()) {
                         if (sh.IsUint64()) {
-                            size_t dim = sh.GetUint64();
+                            int64_t dim = sh.GetUint64();
                             if (dim > std::numeric_limits<dimension_value_t>::max()) {
                                 SPDLOG_WARN("There was an error parsing shape {}", s.name.GetString());
                                 if (!firstErrorStatus.ok()) {
